add layout, split to components
