+++ src/components/Sidebar.js
@@ -1,38 +1,29 @@
 import React from 'react';
 import { withStyles } from '@material-ui/core/styles';
-import titleInitials from "../utils/title-initials";
 import Drawer from '@material-ui/core/Drawer';
 import BottomNavigation from '@material-ui/core/BottomNavigation';
 import BottomNavigationAction from '@material-ui/core/BottomNavigationAction';
-import AddIcon from '@material-ui/icons/Add';
 import ExploreIcon from '@material-ui/icons/Explore';
 import RestoreIcon from '@material-ui/icons/Restore';
 import Divider from '@material-ui/core/Divider';
 import TextField from '@material-ui/core/TextField';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemText from '@material-ui/core/ListItemText';
-import Avatar from '@material-ui/core/Avatar';
-import Button from '@material-ui/core/Button';
+import AddChatButton from './AddChatButton';
+import ChatList from './ChatList';
 
 const styles = theme => ({
 
-  addChatButton: {
-    position: 'absolute',
-    right: '30px',
-    bottom: '60px'
+  textField: {
+    marginLeft: theme.spacing.unit,
+    marginRight: theme.spacing.unit,
+    width: 300,
   },
-
   drawerPaper: {
     position: 'relative',
     width: 320,
     height: '100%',
     overflowY: 'initial'
   },
-
   toolbar: theme.mixins.toolbar,
-
-
 });
 
 const Sidebar = ({classes, chatList}) => (
@@ -51,17 +42,8 @@ const Sidebar = ({classes, chatList}) => (
       />
     </div>
     <Divider />
-    <List className={classes.chatList} component="nav">
-      {chatList.map((chatItem, index) => (
-        <ListItem key={index} button>
-          <Avatar className={classes.avatar}>{titleInitials(chatItem.title)}</Avatar>
-          <ListItemText primary={chatItem.title} secondary={chatItem.date} />
-        </ListItem>
-      ))}
-    </List>
-    <Button variant="fab" color="primary" aria-label="Add" className={classes.addChatButton}>
-      <AddIcon />
-    </Button>
+    <ChatList chatList={chatList}/>
+    <AddChatButton />
     <BottomNavigation showLabels>
       <BottomNavigationAction label="My chats" icon={<RestoreIcon />} />
       <BottomNavigationAction label="Explore" icon={<ExploreIcon />} />
